+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 ## 0.1.9
 
 #### Enhancements
-* Allow setting `protected` in `neon_branch` and `neon_project.branch`
+* Added `protected` in `neon_branch` and `neon_project.branch`
 * Added `neon_connection_uri` datasource
-* Allow setting `history_retention` in `neon_project`
+* Added `history_retention` & `allowed_ips` in `neon_project`
 
 ## 0.1.8